don't allocate for every row of input data
ReadBytes always allocates a copy of the input, ReadSlice returns a reference to the underlying buffer. This allows us to drop the allocations to be a factor of the number/size of keys, instead of the size of the input data. We just need to be careful not to hold onto the ReadSlice longer than the next call to ReadSlice.